+++ src/components/Chat.js
@@ -25,7 +25,9 @@ class Chat extends Component {
 
     this.chatMessage = this.chatMessage.bind(this);
   }
-
+  // shouldComponentUpdate() {
+  //   return false;
+  // }
   chatMessage(name = '', text = '', type = 'info') {
     this.state.messages.push({ name: name, text: text, type: type });
     this.setState({ new_message: true });
