+++ src/components/Worker.js
@@ -23,6 +23,9 @@ class Worker extends Component {
     this.manageAll = this.manageAll.bind(this);
     this.dismiss = this.dismiss.bind(this);
   }
+  // shouldComponentUpdate() {
+  //   return false;
+  // }
   UNSAFE_componentWillMount() {
     let data = this.props.data;
     data.workers.forEach(worker => {
