+++ src/components/content/ProjectsFind.js
@@ -4,7 +4,7 @@ import ProjectOfferBlock from '../ProjectOfferBlock';
 
 import PublicRelations from '../content/PublicRelations';
 import SalesAgency from '../SalesAgency';
-import Bar from '../Bar';
+import Bar from '../Bar/Bar';
 import { colors } from '../../game/knowledge';
 
 class ProjectsFind extends Component {
